[refactor] Move useSelectedLiteGraphItems to composables/canvas
This composable deals with canvas rendering/selection visualization,
not the graph data model, so it belongs in composables/canvas/ instead
of composables/graph/.
